Adjusting data types of key symbols to ensure platform-specific data is always available.

+++ softlight/include/softlight/SL_KeySymWin32.hpp
@@ -9,7 +9,7 @@
 /*-------------------------------------
  * @brief Keyboard Symbols
 -------------------------------------*/
-enum SL_KeySymbol : uint32_t;
+enum SL_KeySymbol : uint16_t;
